#28 move waitlist model to its own app
